code restructured and formatted
