+++ client/src/components/card/description/index.jsx
@@ -1,0 +1,26 @@
+import React from 'react'
+
+import {
+  Wrapper,
+  Container,
+  DescriptionSection,
+} from './style';
+
+const DescriptionCard = ({ description }) => {
+  return (
+    <Wrapper>
+      <Container>
+        {
+          description ?
+          (
+            <DescriptionSection>
+              {description}
+            </DescriptionSection>
+          ) : null
+        }
+      </Container>
+    </Wrapper>
+  )
+}
+
+export default DescriptionCard
