+++ client/src/components/card/description/style.jsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { styleTokens } from "../../../styles/variable";
-import { ellipsis } from "../../../styles/reusableStyle";
+// import { ellipsis } from "../../../styles/reusableStyle";
 
 export const Wrapper = styled.div`
   font-size: 14px;
@@ -19,39 +19,19 @@ export const Container = styled.div`
   padding: 14px;
   font-size: 14px;
   color: ${styleTokens.darkTextColor};
-  width: 90%;
+  width: 100%;
   border: 2px solid ${styleTokens.borderColor};
   border-radius: 12px;
   font-weight: 500;
 `
 
-export const TitleText = styled.div`
-  width: 90%;
-  font-size: 13px;
-  ${ellipsis}
-`
-
-export const Divider = styled.div`
-  height: 2px;
-  width: 95%;
-  background: ${styleTokens.borderColor};
-  margin-top: 12px;
-`
-
 export const DescriptionSection = styled.div`
   font-size: 12px;
-  padding: ${props => props.reducePadding ? '0px' : '12px 0'};
   color: ${styleTokens.lightTextColor};
   font-style: italic;
   font-weight: 400;
   text-align: left;
-  width: 95%;
-  height: 100px;
+  width: 100%;
+  height: 100%;
   opacity: 0.8;
-  line-height: 1.5;
-  overflow: hidden;
-  position: relative;
-   display: -webkit-box;
-  -webkit-line-clamp: 5;
-  -webkit-box-orient: vertical;
 `
