+++ client/src/scenes/wishlist/components/listcontainer/index.jsx
@@ -5,43 +5,46 @@ import { getDetailedListCount } from '../../../../helper';
 
 import Button from '../../../../components/button';
 import WishCard from '../../../../components/card/wish';
+import DescriptionCard from '../../../../components/card/description';
 
 import {
   IconWrapper,
   CheckListIcon,
-  DateIcon
+  DateIcon,
+  CheckedIcon,
+  UncheckedIcon
 } from '../../../../styles/icon';
 
 import{
   WishWrapper,
-  ThirdSlot,
+  Slot,
   Row,
   ListSection,
   TitleText,
   DateTimeText,
   ListCount,
   SubText,
+  TaskInfoText
 } from './style';
 
-const ListContainer = (wishlistData) => {
+const ListContainer = ({wishlistData}) => {
   const {
     isCompleted,
     createdDate,
     hasBucket,
-    list = []
+    list = [],
+    description
   } = wishlistData;
-  return (
-    <ListSection>
-      <Row>
-        <TitleText>
-          Your Wishes
-        </TitleText>
-        {hasBucket ?
+
+  const getListCount = () => (
+    <>
+      {hasBucket ?
         (
           <ListCount isCompleted={isCompleted}>
             <IconWrapper
               iconColor={isCompleted ?
-                styleTokens.lightActiveColor : styleTokens.lightIconColor
+                styleTokens.lightActiveColor :
+                styleTokens.lightIconColor
               }
               iconSize={16}
             >
@@ -52,9 +55,45 @@ const ListContainer = (wishlistData) => {
             </SubText>
           </ListCount>
         ) : null
-        }
+      }
+    </>
+  )
+
+  const getCompletionStatus = () => {
+    const taskInfoText = isCompleted ? "completed" : "not completed yet";
+    return(
+      <Slot>
+        <Row>
+          <IconWrapper
+          iconColor={
+            isCompleted ?
+            styleTokens.lightActiveColor :
+            styleTokens.lightIconColor
+          }
+          iconSize={14}
+        >
+          {isCompleted ? <CheckedIcon /> : <UncheckedIcon />}
+        </IconWrapper>
+        <TaskInfoText isCompleted={isCompleted}>
+          {hasBucket ?
+            `All task are ${taskInfoText}` :
+            `This task is ${taskInfoText}`
+          }
+        </TaskInfoText>
       </Row>
-      <ThirdSlot>
+    </Slot>
+    )
+  }
+
+  return (
+    <ListSection>
+      <Row>
+        <TitleText>
+          Your Wishes
+        </TitleText>
+        {getListCount()}
+      </Row>
+      <Slot>
         <Row>
           <IconWrapper
             iconColor={ styleTokens.lightIconColor}
@@ -66,18 +105,19 @@ const ListContainer = (wishlistData) => {
             {createdDate}
           </DateTimeText>
         </Row>
-      </ThirdSlot>
+        {getCompletionStatus()}
+      </Slot>
       <WishWrapper>
-        <>
+        {hasBucket ? (
+          <>
           {list.map((data, idx)=>(
             <WishCard {...data} key={idx}/>
           ))}
         </>
+        ) : <DescriptionCard description={description} /> }
         < Button/>
       </WishWrapper>
     </ListSection>
-
-
   )
 }
