+++ client/src/scenes/wishlist/index.jsx
@@ -1,0 +1,39 @@
+import React from 'react';
+import { withRouter } from "react-router";
+import queryString from 'query-string';
+
+import { getWishlistById } from '../../helper';
+import { isEmptyObject } from '../../utils';
+
+import WishDetailsView from '../../scenes/wishlist/components/wishdetails';
+
+const Wishlist = (props) => {
+  const [wishlist, updateWishlist] = React.useState(null);
+  const { history, location: { search } } = props;
+  const { id } = queryString.parse(search);
+
+  React.useEffect(()=>{
+    const newWishlist = getWishlistById(id);
+    updateWishlist(newWishlist);
+  }, [id]);
+
+  return (
+    <>
+      {wishlist === null ? (
+        <>Loading Data</>
+      ): (
+        <>
+          {
+            !isEmptyObject(wishlist) ?
+            (
+              <WishDetailsView wishlist={wishlist} />
+            ) : (<>No Data Found</>)
+          }
+        </>
+      )
+      }
+    </>
+  )
+};
+
+export default withRouter(Wishlist);
