+++ client/src/styles/icon.js
@@ -7,8 +7,9 @@ export { default as FavoriteMarkedIcon } from "@material-ui/icons/FavoriteRounde
 export { default as CheckListIcon } from "@material-ui/icons/PlaylistAddCheckRounded";
 export { default as DateIcon } from "@material-ui/icons/EventAvailable";
 export { default as CheckedIcon } from '@material-ui/icons/CheckCircleOutline';
-export {default as UncheckedIcon } from '@material-ui/icons/RadioButtonUnchecked';
-export {default as DescriptionIcon } from '@material-ui/icons/DescriptionOutlined';
+export { default as UncheckedIcon } from '@material-ui/icons/RadioButtonUnchecked';
+export { default as DescriptionIcon } from '@material-ui/icons/DescriptionOutlined';
+export { default as TagIcon } from '@material-ui/icons/LocalOfferRounded';
 
 export const IconWrapper = styled.span`
   padding-right: ${(props) => (props.pr !== undefined ? `${props.pr}px` : "8px")};
